+++ packages/bitcoin/src/datasource/indexer/indexer.data-source.ts
@@ -10,19 +10,62 @@ import {
   Balance,
   FeeData,
   DefaultFeeOptions,
+  TransactionData,
+  TransactionStatus,
 } from '@xdefi-tech/chains-core';
-import { UTXOManifest } from '@xdefi-tech/chains-utxo';
+import * as Bitcoin from 'bitcoinjs-lib';
+import { UTXO, UTXOManifest } from '@xdefi-tech/chains-utxo';
 import { utils } from 'ethers';
 import { Observable } from 'rxjs';
 
 import { ChainMsg } from '../../msg';
 
-import { getBalance, getTransactions, getFees, getNFTBalance } from './queries';
+import {
+  getBalance,
+  getTransactions,
+  getFees,
+  getNFTBalance,
+  broadcast,
+  scanUTXOs,
+  getTransactionByHash,
+} from './queries';
 
 @Injectable()
 export class IndexerDataSource extends DataSource {
   constructor(manifest: UTXOManifest) {
     super(manifest);
+  }
+
+  async broadcast(messages: ChainMsg[]): Promise<Transaction[]> {
+    const result: Transaction[] = [];
+    for await (const message of messages) {
+      const { signedTransaction } = message;
+
+      if (!message.hasSignature || !signedTransaction) {
+        throw new Error(`Message ${JSON.stringify(message)} is not signed`);
+      }
+
+      const txHash = await broadcast(signedTransaction);
+
+      result.push(Transaction.fromData({ hash: txHash }));
+    }
+
+    return result;
+  }
+
+  async scanUTXOs(address: string): Promise<UTXO[]> {
+    const utxos = await scanUTXOs(address);
+    return utxos.map((utxo) => {
+      const tx = Bitcoin.Transaction.fromHex(utxo.oTxHex as string);
+      const utxoMapped: UTXO = {
+        hash: utxo.oTxHash,
+        index: utxo.oIndex,
+        value: parseInt(utxo.value.value),
+        txHex: utxo.oTxHex as string,
+        witnessUtxo: tx.outs[utxo.oIndex],
+      };
+      return utxoMapped;
+    });
   }
 
   async getNFTBalance(address: string) {
@@ -95,5 +138,24 @@ export class IndexerDataSource extends DataSource {
 
   async getNonce(_address: string): Promise<number> {
     throw new Error('Method not implemented.');
+  }
+
+  async getTransaction(txHash: string): Promise<TransactionData | null> {
+    const tx = await getTransactionByHash(txHash);
+    let response = null;
+
+    if (tx && tx.hash) {
+      response = {
+        hash: tx.hash,
+        from: (tx.inputs && tx.inputs[0].address) || '',
+        to: (tx.outputs && tx.outputs[0].address) || '',
+        status:
+          (tx.blockNumber || 0) > 0
+            ? TransactionStatus.success
+            : TransactionStatus.pending,
+      };
+    }
+
+    return response;
   }
 }