+++ packages/bitcoin/src/signers/private-key.signer.ts
@@ -2,6 +2,7 @@
 import { Signer, SignerDecorator } from '@xdefi-tech/chains-core';
 import { secp256k1 } from '@noble/curves/secp256k1';
 import * as btc from '@scure/btc-signer';
+import * as Bitcoin from 'bitcoinjs-lib';
 
 import { ChainMsg } from '../msg';
 
@@ -62,6 +63,20 @@ export class PrivateKeySigner extends Signer.Provider {
     txP2WPKH.finalize();
 
     message.sign(txP2WPKH.hex);
+  }
+
+  async signRawTransaction(
+    txHex: string,
+    derivation?: string
+  ): Promise<string> {
+    const pk = Bitcoin.ECPair.fromWIF(
+      await this.getPrivateKey(derivation ?? '')
+    );
+    const psbt = Bitcoin.Psbt.fromHex(txHex);
+
+    psbt.signAllInputs(pk);
+    psbt.finalizeAllInputs();
+    return psbt.extractTransaction(true).toHex();
   }
 }