+++ packages/litecoin/src/signers/seed-phrase.signer.spec.ts
@@ -50,6 +50,16 @@ describe('seed-phrase.signer', () => {
     expect(message.signedTransaction).toBeTruthy();
   });
 
+  it('should sign a raw transaction using a seed phrase', async () => {
+    const txHex =
+      '70736274ff0100710200000001467066a38eef3daa8a2bb0e35dccfd098e505b17efa7fcee45e390446679c35c0000000000ffffffff0264000000000000001600145c62b57d49ea96f0ff6a8473c3332ec3fc501ca4ca250000000000001600145c62b57d49ea96f0ff6a8473c3332ec3fc501ca4000000000001011f10270000000000001600145c62b57d49ea96f0ff6a8473c3332ec3fc501ca4000000';
+    const signedTx = await signer.signRawTransaction(txHex, derivation);
+
+    expect(signedTx).toEqual(
+      '02000000000101467066a38eef3daa8a2bb0e35dccfd098e505b17efa7fcee45e390446679c35c0000000000ffffffff0264000000000000001600145c62b57d49ea96f0ff6a8473c3332ec3fc501ca4ca250000000000001600145c62b57d49ea96f0ff6a8473c3332ec3fc501ca40248304502210095104f03cf04aa7e9f8516b9280dc30ad091f7a1aafb49e9e1bdd8c4513701f2022039e449e17cc4613128b320b15c6a7f2937158f0fa6b6714f8472c6f035d977c00121026d165716c33a95fbbed7c3bdfc155afc43dce53eacf059fc65aa6092f8625ad800000000'
+    );
+  });
+
   it('should return false when verifing an invalid address', async () => {
     expect(signer.verifyAddress('0xDEADBEEF')).toBe(false);
   });