+++ packages/cosmos/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @xdefi-tech/chains-cosmos
 
+## 2.0.10
+
+### Patch Changes
+
+- 3b42f94: Feat: temporary remove osmojs lib
+
+## 2.0.9
+
+### Patch Changes
+
+- 62105c4: Feat: update README.md
+
 ## 2.0.8
 
 ### Patch Changes