+++ packages/solana/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @xdefi-tech/chains-solana
 
+## 2.0.10
+
+### Patch Changes
+
+- 62105c4: Feat: update README.md
+
 ## 2.0.9
 
 ### Patch Changes